requirements.txt
netaddr==0.9.0
nmcli
pyImprov @ git+https://github.com/Mimoja/pyImprov@main
